+++ app/routes/components/Modal/changelog.codegen.ts
@@ -3,4 +3,6 @@
 
 export const changelog = ''
   + '- 2025-10-13: Initial version.\n'
+  + '- 2025-10-14: Make `Modal.Content` support `center` and `shadow` props. Update\n'
+  + '  `Modal.Content` styles.\n'
   ;