+++ src/components/Modal/index.ts
@@ -1,3 +1,3 @@
-// Modal from copy-ui @ 2025-10-13.
+// Modal from copy-ui @ 2025-10-14.
 
 export { Modal } from "./Modal";