+++ src/components/Modal/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-10-13: Initial version.
+- 2025-10-14: Make `Modal.Content` support `center` and `shadow` props. Update
+  `Modal.Content` styles.